+++ test/OpenTelemetry.AutoInstrumentation.Tests/Configurations/SettingsTests.cs
@@ -250,6 +250,7 @@ public class SettingsTests : IDisposable
     [InlineData("WCFCLIENT", TracerInstrumentation.WcfClient)]
 #endif
     [InlineData("MYSQLCONNECTOR", TracerInstrumentation.MySqlConnector)]
+    [InlineData("AZURE", TracerInstrumentation.Azure)]
     internal void TracerSettings_Instrumentations_SupportedValues(string tracerInstrumentation, TracerInstrumentation expectedTracerInstrumentation)
     {
         Environment.SetEnvironmentVariable(ConfigurationKeys.Traces.TracesInstrumentationEnabled, "false");